MDEXP-630 - code refactoring, update test
