+++ src/main/java/org/folio/dataexp/exception/file/definition/FileExtensionException.java
@@ -1,4 +1,4 @@
-package org.folio.dataexp.exception.export;
+package org.folio.dataexp.exception.file.definition;
 
 public class FileExtensionException extends RuntimeException {
   public FileExtensionException(String message) {
