+++ src/main/java/org/folio/dataexp/exception/file/definition/FileSizeException.java
@@ -1,4 +1,4 @@
-package org.folio.dataexp.exception.export;
+package org.folio.dataexp.exception.file.definition;
 
 public class FileSizeException extends RuntimeException{
   public FileSizeException(String message) {
