+++ src/main/java/org/folio/dataexp/exception/file/definition/UploadFileException.java
@@ -1,4 +1,4 @@
-package org.folio.dataexp.exception.export;
+package org.folio.dataexp.exception.file.definition;
 
 public class UploadFileException extends RuntimeException {
   public UploadFileException (String message) {
