+++ src/test/java/org/folio/dataexp/controllers/FileDefinitionsControllerTest.java
@@ -5,7 +5,7 @@ import org.folio.dataexp.BaseTest;
 import org.folio.dataexp.domain.dto.FileDefinition;
 import org.folio.dataexp.domain.entity.FileDefinitionEntity;
 import org.folio.dataexp.repository.FileDefinitionEntityRepository;
-import org.folio.dataexp.service.DataExportService;
+import org.folio.dataexp.service.FileDefinitionsService;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
@@ -18,12 +18,12 @@ import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.http.MediaType.APPLICATION_OCTET_STREAM_VALUE;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class DataExportControllerTest extends BaseTest {
+public class FileDefinitionsControllerTest extends BaseTest {
 
   @MockBean
   private FileDefinitionEntityRepository fileDefinitionEntityRepository;
   @MockBean
-  private DataExportService dataExportService;
+  private FileDefinitionsService fileDefinitionsService;
 
   @Test
   @SneakyThrows
@@ -33,7 +33,7 @@ public class DataExportControllerTest extends BaseTest {
     fileDefinition.setId(UUID.randomUUID());
     fileDefinition.setUploadFormat(FileDefinition.UploadFormatEnum.CSV);
 
-    when(dataExportService.postFileDefinition(isA(FileDefinition.class))).thenReturn(fileDefinition);
+    when(fileDefinitionsService.postFileDefinition(isA(FileDefinition.class))).thenReturn(fileDefinition);
 
     mockMvc.perform(MockMvcRequestBuilders
         .post("/data-export/file-definitions")
